db_admin/pyproject.toml
[tool.poetry]
name = "db_admin"
version = "0.1.0"
description = ""
authors = ["Ez Ablaza <ezra.ablaza.inbox@gmail.com>"]
readme = "README.md"
packages = [{include = "db_admin"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
dj-database-url = "^2.0.0"
psycopg = "^3.1.10"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
